requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
redis[hiredis]>=5.2.0
structlog>=24.4.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
python-dotenv>=1.0.1
asyncpg>=0.30.0
yfinance>=0.2.40
edgartools>=3.0.0
openai>=1.50.0
tiktoken>=0.12.0
chromadb>=1.5.0
openpyxl>=3.1.0
python-pptx>=1.0.0
PyJWT[crypto]>=2.8.0
